Singleton metaclass was relocated from FrequencyTable.py to Window.py
