+++ lab1/Window.py
@@ -1,5 +1,14 @@
 import tkinter as tk
 from tkinter import ttk
+
+class Singleton(type):
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
+
 class BaseWindow:
     def __init__(self,data):
         self.__root = tk.Tk()
